runner: move rungroup into cmd/epp/runner package
